Fix: V823 Decreased performance. Object may be created in-place in the 'allPoints' container. Consider replacing methods: 'push_back' -> 'emplace_back'. world TransporterHandler.cpp 218
V823 Decreased performance. Object may be created in-place in the 'splinePath' container. Consider replacing methods: 'push_back' -> 'emplace_back'. world TransporterHandler.cpp 248
V220 Suspicious sequence of types castings: memsize -> 32-bit integer -> memsize. The value being cast: 'm_players[Team].size()'. world Battleground.cpp 1057

+++ src/world/Management/Battleground/Battleground.cpp
@@ -1054,7 +1054,7 @@ bool CBattleground::HasFreeSlots(uint32 Team, uint32 type)
 
     if (isArena(type))
     {
-        res = (static_cast<uint32>(m_players[Team].size()) + m_pendPlayers[Team].size() < maxPlayers);
+        res = m_players[Team].size() + m_pendPlayers[Team].size() < maxPlayers;
     }
     else
     {
